refactor: split erc20 from uniswap
